set more buetiful start values for noise

+++ middleware/src/generator/texture/chess_cubes.cpp
@@ -5,7 +5,7 @@
 
 
 math::Generator3D ChessCubes::Result() const {
-    return math::Generator3D([frequency = m_frequency](double x, double y, double z) -> double {
+    return math::Generator3D([frequency = static_cast<double>(m_frequency)](double x, double y, double z) -> double {
         auto ix = math::ToInt32Continuous(x * frequency);
         auto iy = math::ToInt32Continuous(y * frequency);
         auto iz = math::ToInt32Continuous(z * frequency);
